Pod update
Pod files where updated. Now we have a new pod file that will be used for the creation of charts. If this update causes any conflicts with the current functionality, we will have to check that all the pod names are included in the pod file and then update again.